+++ query.js
@@ -1,0 +1,7 @@
+"use strict";
+class Query {
+    constructor(db) {
+        this.db = db;
+    }
+}
+module.exports = Query;
