+++ query.ts
@@ -1,0 +1,9 @@
+class Query
+{
+  constructor(db)
+  {
+    this.db = db;
+  }
+}
+
+module.exports = Query;
